Save favorite character when stars click into datails screen
