+++ android/MarvelChallenge/app/src/main/java/com/manoelsrs/marvelchallenge/di/module/CharactersFragmentModule.kt
@@ -3,7 +3,6 @@ package com.manoelsrs.marvelchallenge.di.module
 import com.manoelsrs.marvelchallenge.di.PerFragment
 import com.manoelsrs.marvelchallenge.presentation.home.characters.actions.GetCharacters
 import com.manoelsrs.marvelchallenge.presentation.home.characters.actions.GetOffset
-import com.manoelsrs.marvelchallenge.presentation.home.characters.actions.SaveFavorite
 import com.manoelsrs.marvelchallenge.presentation.home.characters.fragment.viewmodel.CharactersViewModel
 import com.manoelsrs.marvelchallenge.repository.Repository
 import dagger.Module
@@ -26,13 +25,8 @@ class CharactersFragmentModule {
 
     @PerFragment
     @Provides
-    fun providesSaveFavoriteAction(repository: Repository) =
-        SaveFavorite(repository)
-
-    @PerFragment
-    @Provides
-    fun providesCharacterViewModel(getCharacters: GetCharacters, saveFavorite: SaveFavorite) =
-        CharactersViewModel(getCharacters, saveFavorite)
+    fun providesCharacterViewModel(getCharacters: GetCharacters) =
+        CharactersViewModel(getCharacters)
 
     @PerFragment
     @Provides
