+++ android/MarvelChallenge/app/src/main/java/com/manoelsrs/marvelchallenge/presentation/home/details/DetailsActivity.kt
@@ -44,7 +44,7 @@ class DetailsActivity : BaseActivity(), DetailsContract {
             override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {
                 super.onScrollStateChanged(recyclerView, newState)
                 if (!rvSeries.canScrollHorizontally(1))
-                    presenter.loadMoreComics()
+                    presenter.loadMoreSeries()
             }
         })
     }
